+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - changed : icon "partie de" use entity color
 - changed : evolution update date use futur and past color for variable icon
 - changed : put tab title on two rows
+- changed : make tab title percent use entity color
 - fixed : jsonjsdv_editor evolution dont stringify null value
 
 ## 0.10.4 (2025-01-26)